refactor : GlobalStyle 컴포넌트 파일 명칭 변경
